Upgrade RN to 29 (NB: You may need to upgrade some packages in the Android SDK Manager to make it to run on Android)
I got a problem with a support library dependency, which i solved by installing all upgrades in the Android SDK Manager
