+++ android/app/src/main/java/com/reactnativeworkshop/MainActivity.java
@@ -1,10 +1,6 @@
 package com.reactnativeworkshop;
 
 import com.facebook.react.ReactActivity;
-import com.github.xinthink.rnmk.ReactMaterialKitPackage;
-import com.imagepicker.ImagePickerPackage;
-import com.facebook.react.ReactPackage;
-import com.facebook.react.shell.MainReactPackage;
 
 import java.util.Arrays;
 import java.util.List;
@@ -18,27 +14,5 @@ public class MainActivity extends ReactActivity {
     @Override
     protected String getMainComponentName() {
         return "ReactNativeWorkshop";
-    }
-
-    /**
-     * Returns whether dev mode should be enabled.
-     * This enables e.g. the dev menu.
-     */
-    @Override
-    protected boolean getUseDeveloperSupport() {
-        return BuildConfig.DEBUG;
-    }
-
-    /**
-     * A list of packages used by the app. If the app uses additional views
-     * or modules besides the default ones, add more packages here.
-     */
-    @Override
-    protected List<ReactPackage> getPackages() {
-        return Arrays.<ReactPackage>asList(
-            new MainReactPackage(),
-        new ReactMaterialKitPackage(),
-        new ImagePickerPackage()
-        );
     }
 }
